Move most of lorax into pylorax and add sphinx-argparse
This lets us use sphinx-argparse to document the cmdline args for lorax.
It needs to be importable from a module, not a script, so move it into
pylorax.lorax

NOTE: This needs sphinx-argparse to be packaged and have py3 support
before it will really work. Should also use my mock method of running
sphinx (see anaconda) instead of venv and pip modules.