EXPERIMENTAL: Multiple release repo support
This is just an experiment, to see if we can create multiple yum objects
and select them for depsolving and image creation.

In my limited testing this does work, at least for 7.6 and 7.7 content.

But there's no guarantee that it would work for all 7.x, and AFAIK we do
not make any promises to that effect so while it may work across
releases, it isn't on purpose.